Added Lots (See description for details)
Added red screen flash on damage taken.

Added damage numbers. (Feel free to change, kind of looks bad with this font.)

Added scaling to the health and damage of enemies based on loops and a multiplier. (We can decide on actual equations during a meeting)

Added Invincibility Frames. (Time of invincibility is changeable on the player)